usart: rename and rearrange usart types, namely StopBits, Parity, DataBits, Oversampling and Config
Derive or implement necessary traits for those types.
